# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estore API",
        "description": "Estore Api",
        "contact": {
            "name": "Victor Maamaa",
            "email": "maaamaavictor@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/brands": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Get a list of brands",
                "operationId": "130f2bd423b6fe2355253e2c39878d99",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (default: 10)",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (default: 1)",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by brand name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_visible",
                        "in": "query",
                        "description": "Filter by visibility (true/false)",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Brands retrieved successfully",
                                    "data": {
                                        "current_page": 1,
                                        "data": [
                                            {
                                                "id": 1,
                                                "name": "Lockman Ltd",
                                                "slug": "lockman-ltd",
                                                "website": "https://www.kerluke.com",
                                                "description": "Bill,' thought Alice,) 'Well, I shan't grow any more--As it is, I suppose?' 'Yes,' said Alice sharply, for she had drunk half the bottle, she found she could see, as she could not even room for her.",
                                                "position": 0,
                                                "is_visible": true,
                                                "seo_title": null,
                                                "seo_description": null,
                                                "sort": null,
                                                "created_at": "2023-06-09T14:02:19.000000Z",
                                                "updated_at": "2023-08-09T22:31:03.000000Z"
                                            }
                                        ],
                                        "first_page_url": "...",
                                        "from": 1,
                                        "last_page": 1,
                                        "0": "... (pagination details)"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Error message",
                                    "errors": "{...} Validation errors or other errors"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Brands"
                ],
                "summary": "Create a new brand",
                "operationId": "283de011874bd2a1ada6efafb676e83f",
                "requestBody": {
                    "description": "Brand creation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "name": "Brand Name",
                                "slug": "brand-slug",
                                "website": "https://www.brand.com",
                                "description": "Brand Description",
                                "position": 0,
                                "is_visible": true,
                                "seo_title": "SEO Title",
                                "seo_description": "SEO Description",
                                "sort": 1
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Brand created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Brand created successfully",
                                    "data": {
                                        "id": 1,
                                        "name": "Brand Name",
                                        "slug": "brand-slug",
                                        "website": "https://www.brand.com",
                                        "description": "Brand Description",
                                        "position": 0,
                                        "is_visible": true,
                                        "seo_title": "SEO Title",
                                        "seo_description": "SEO Description",
                                        "sort": 1,
                                        "created_at": "2024-01-08T13:14:29.000000Z",
                                        "updated_at": "2024-01-08T13:14:29.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Validation error",
                                    "errors": {
                                        "name": [
                                            "The name field is required."
                                        ],
                                        "0": " // ... (other validation errors)"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}